requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-formtools==2.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
kavenegar==1.1.2
numpy==1.23.1
openpyxl==3.0.10
Pillow==9.2.0
pycryptodome==3.15.0
pypep-pepco==0.1.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
xlrd==1.2.0
XlsxWriter==3.0.3
xmltodict==0.13.0
